🎨 Move make_compound into core
